backend/requirements.txt
Flask
flask-cors
groq
python-dotenv
pymongo
reportlab
pymupdf
python-docx
gunicorn
numpy==1.26.4

